ml/requirements.txt
# Python 3.10.19 Compatible Requirements
# Core FastAPI dependencies
fastapi>=0.100.0,<0.110.0
uvicorn>=0.23.0,<0.25.0
pydantic>=2.0.0,<2.10.0
pydantic_core>=2.0.0,<2.40.0
starlette>=0.27.0,<0.38.0

# AI and API integrations
mistralai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<1.1.0

# Web scraping
scrapingbee>=2.0.0,<3.0.0
beautifulsoup4>=4.12.0,<4.13.0
lxml>=4.9.0,<5.0.0

# Web server
h11>=0.14.0,<0.15.0
anyio>=3.7.0,<4.0.0

# NLP and Text Processing
spacy>=3.5.0,<3.8.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
textblob>=0.17.0,<0.18.0
nltk>=3.8.0,<3.9.0

# Spacy dependencies
thinc>=8.1.0,<8.3.0
spacy-legacy>=3.0.12,<3.1.0
spacy-loggers>=1.0.4,<1.1.0
catalogue>=2.0.8,<2.1.0
cymem>=2.0.6,<2.1.0
preshed>=3.0.8,<3.1.0
murmurhash>=1.0.9,<1.1.0
wasabi>=1.1.0,<1.2.0
srsly>=2.4.0,<2.5.0
blis>=0.7.9,<1.1.0
confection>=0.1.0,<0.2.0
typer-slim>=0.9.0,<0.13.0
cloudpathlib>=0.15.0,<0.17.0
smart_open>=6.3.0,<7.0.0
weasel>=0.3.0,<0.4.0

# Data processing
numpy>=1.24.0,<1.27.0
joblib>=1.3.0,<1.4.0
tqdm>=4.65.0,<4.67.0
regex>=2023.10.0,<2024.0.0

# HTTP and networking
requests>=2.31.0,<2.33.0
urllib3>=2.0.0,<2.3.0
certifi>=2023.7.0
charset-normalizer>=3.0.0,<3.4.0
idna>=3.4,<3.10.0

# Utilities
click>=8.1.0,<8.2.0
Jinja2>=3.1.0,<3.2.0
MarkupSafe>=2.1.0,<3.0.0
packaging>=23.0,<25.0
typing_extensions>=4.7.0,<4.13.0
typing-inspection>=0.4.0,<0.5.0
setuptools>=68.0.0,<75.0.0
wrapt>=1.15.0,<1.17.0

# PDF processing
pypdf>=3.15.0,<5.0.0

# Audio (optional - only if needed)
# pydub>=0.25.1,<0.26.0
# SpeechRecognition>=3.10.0,<3.12.0

# Mistral AI and environment
mistralai>=0.4.0,<2.0.0
python-dotenv>=1.0.0,<1.1.0

# Additional utilities
annotated-types>=0.5.0,<0.8.0
